Buying basic improvements works correctly

+++ WPF App/WPF App/Model/User.cs
@@ -8,12 +8,12 @@ namespace WPF_App.Model
 {
     public class User
     {
-        private static int _points;
-        private static int _speedOfAddingPoints;
+        private static double _points;
+        private static double _speedOfAddingPoints;
         private static int _level;
 
-        public int Points { get => _points; set => _points = value; }
-        public int SpeedOfAddingPoints { get => _speedOfAddingPoints; set => _speedOfAddingPoints = value; }
+        public double Points { get => _points; set => _points = value; }
+        public double SpeedOfAddingPoints { get => _speedOfAddingPoints; set => _speedOfAddingPoints = value; }
         public int Level { get => _level; set => _level = value; }
 
         // CONSTRUCTORS
